backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic[email]==2.5.3
openai==1.10.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
email-validator==2.1.0
itsdangerous==2.1.2
